# Conflicts:
#	TMessagesProj/build.gradle
#	TMessagesProj/src/main/java/org/telegram/messenger/BuildVars.java
#	gradle.properties

+++ TMessagesProj/src/main/java/com/google/android/exoplayer2/upstream/cache/CachedContent.java
@@ -16,34 +16,40 @@
 package com.google.android.exoplayer2.upstream.cache;
 
 import static com.google.android.exoplayer2.util.Assertions.checkArgument;
+import static com.google.android.exoplayer2.util.Assertions.checkNotNull;
 import static com.google.android.exoplayer2.util.Assertions.checkState;
+import static java.lang.Math.max;
+import static java.lang.Math.min;
 
 import androidx.annotation.Nullable;
+import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.util.Log;
 import java.io.File;
+import java.util.ArrayList;
 import java.util.TreeSet;
 
-/** Defines the cached content for a single stream. */
+/** Defines the cached content for a single resource. */
 /* package */ final class CachedContent {
 
   private static final String TAG = "CachedContent";
 
-  /** The cache file id that uniquely identifies the original stream. */
+  /** The cache id that uniquely identifies the resource. */
   public final int id;
-  /** The cache key that uniquely identifies the original stream. */
+  /** The cache key that uniquely identifies the resource. */
   public final String key;
   /** The cached spans of this content. */
   private final TreeSet<SimpleCacheSpan> cachedSpans;
+  /** Currently locked ranges. */
+  private final ArrayList<Range> lockedRanges;
+
   /** Metadata values. */
   private DefaultContentMetadata metadata;
-  /** Whether the content is locked. */
-  private boolean locked;
 
   /**
    * Creates a CachedContent.
    *
-   * @param id The cache file id.
-   * @param key The cache stream key.
+   * @param id The cache id of the resource.
+   * @param key The cache key of the resource.
    */
   public CachedContent(int id, String key) {
     this(id, key, DefaultContentMetadata.EMPTY);
@@ -53,7 +59,8 @@ import java.util.TreeSet;
     this.id = id;
     this.key = key;
     this.metadata = metadata;
-    this.cachedSpans = new TreeSet<>();
+    cachedSpans = new TreeSet<>();
+    lockedRanges = new ArrayList<>();
   }
 
   /** Returns the metadata. */
@@ -72,14 +79,58 @@ import java.util.TreeSet;
     return !metadata.equals(oldMetadata);
   }
 
-  /** Returns whether the content is locked. */
-  public boolean isLocked() {
-    return locked;
+  /** Returns whether the entire resource is fully unlocked. */
+  public boolean isFullyUnlocked() {
+    return lockedRanges.isEmpty();
   }
 
-  /** Sets the locked state of the content. */
-  public void setLocked(boolean locked) {
-    this.locked = locked;
+  /**
+   * Returns whether the specified range of the resource is fully locked by a single lock.
+   *
+   * @param position The position of the range.
+   * @param length The length of the range, or {@link C#LENGTH_UNSET} if unbounded.
+   * @return Whether the range is fully locked by a single lock.
+   */
+  public boolean isFullyLocked(long position, long length) {
+    for (int i = 0; i < lockedRanges.size(); i++) {
+      if (lockedRanges.get(i).contains(position, length)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /**
+   * Attempts to lock the specified range of the resource.
+   *
+   * @param position The position of the range.
+   * @param length The length of the range, or {@link C#LENGTH_UNSET} if unbounded.
+   * @return Whether the range was successfully locked.
+   */
+  public boolean lockRange(long position, long length) {
+    for (int i = 0; i < lockedRanges.size(); i++) {
+      if (lockedRanges.get(i).intersects(position, length)) {
+        return false;
+      }
+    }
+    lockedRanges.add(new Range(position, length));
+    return true;
+  }
+
+  /**
+   * Unlocks the currently locked range starting at the specified position.
+   *
+   * @param position The starting position of the locked range.
+   * @throws IllegalStateException If there was no locked range starting at the specified position.
+   */
+  public void unlockRange(long position) {
+    for (int i = 0; i < lockedRanges.size(); i++) {
+      if (lockedRanges.get(i).position == position) {
+        lockedRanges.remove(i);
+        return;
+      }
+    }
+    throw new IllegalStateException();
   }
 
   /** Adds the given {@link SimpleCacheSpan} which contains a part of the content. */
@@ -93,36 +144,45 @@ import java.util.TreeSet;
   }
 
   /**
-   * Returns the span containing the position. If there isn't one, it returns a hole span
-   * which defines the maximum extents of the hole in the cache.
+   * Returns the cache span corresponding to the provided range. See {@link
+   * Cache#startReadWrite(String, long, long)} for detailed descriptions of the returned spans.
+   *
+   * @param position The position of the span being requested.
+   * @param length The length of the span, or {@link C#LENGTH_UNSET} if unbounded.
+   * @return The corresponding cache {@link SimpleCacheSpan}.
    */
-  public SimpleCacheSpan getSpan(long position) {
+  public SimpleCacheSpan getSpan(long position, long length) {
     SimpleCacheSpan lookupSpan = SimpleCacheSpan.createLookup(key, position);
     SimpleCacheSpan floorSpan = cachedSpans.floor(lookupSpan);
     if (floorSpan != null && floorSpan.position + floorSpan.length > position) {
       return floorSpan;
     }
     SimpleCacheSpan ceilSpan = cachedSpans.ceiling(lookupSpan);
-    return ceilSpan == null ? SimpleCacheSpan.createOpenHole(key, position)
-        : SimpleCacheSpan.createClosedHole(key, position, ceilSpan.position - position);
+    if (ceilSpan != null) {
+      long holeLength = ceilSpan.position - position;
+      length = length == C.LENGTH_UNSET ? holeLength : min(holeLength, length);
+    }
+    return SimpleCacheSpan.createHole(key, position, length);
   }
 
   /**
-   * Returns the length of the cached data block starting from the {@code position} to the block end
-   * up to {@code length} bytes. If the {@code position} isn't cached then -(the length of the gap
-   * to the next cached data up to {@code length} bytes) is returned.
+   * Returns the length of continuously cached data starting from {@code position}, up to a maximum
+   * of {@code maxLength}. If {@code position} isn't cached, then {@code -holeLength} is returned,
+   * where {@code holeLength} is the length of continuously un-cached data starting from {@code
+   * position}, up to a maximum of {@code maxLength}.
    *
    * @param position The starting position of the data.
-   * @param length The maximum length of the data to be returned.
-   * @return the length of the cached or not cached data block length.
+   * @param length The maximum length of the data or hole to be returned.
+   * @return The length of continuously cached data, or {@code -holeLength} if {@code position}
+   *     isn't cached.
    */
   public long getCachedBytesLength(long position, long length) {
     checkArgument(position >= 0);
     checkArgument(length >= 0);
-    SimpleCacheSpan span = getSpan(position);
+    SimpleCacheSpan span = getSpan(position, length);
     if (span.isHoleSpan()) {
       // We don't have a span covering the start of the queried region.
-      return -Math.min(span.isOpenEnded() ? Long.MAX_VALUE : span.length, length);
+      return -min(span.isOpenEnded() ? Long.MAX_VALUE : span.length, length);
     }
     long queryEndPosition = position + length;
     if (queryEndPosition < 0) {
@@ -138,14 +198,14 @@ import java.util.TreeSet;
         }
         // We expect currentEndPosition to always equal (next.position + next.length), but
         // perform a max check anyway to guard against the existence of overlapping spans.
-        currentEndPosition = Math.max(currentEndPosition, next.position + next.length);
+        currentEndPosition = max(currentEndPosition, next.position + next.length);
         if (currentEndPosition >= queryEndPosition) {
           // We've found spans covering the queried region.
           break;
         }
       }
     }
-    return Math.min(currentEndPosition - position, length);
+    return min(currentEndPosition - position, length);
   }
 
   /**
@@ -160,9 +220,9 @@ import java.util.TreeSet;
   public SimpleCacheSpan setLastTouchTimestamp(
       SimpleCacheSpan cacheSpan, long lastTouchTimestamp, boolean updateFile) {
     checkState(cachedSpans.remove(cacheSpan));
-    File file = cacheSpan.file;
+    File file = checkNotNull(cacheSpan.file);
     if (updateFile) {
-      File directory = file.getParentFile();
+      File directory = checkNotNull(file.getParentFile());
       long position = cacheSpan.position;
       File newFile = SimpleCacheSpan.getCacheFile(directory, id, position, lastTouchTimestamp);
       if (file.renameTo(newFile)) {
@@ -185,7 +245,9 @@ import java.util.TreeSet;
   /** Removes the given span from cache. */
   public boolean removeSpan(CacheSpan span) {
     if (cachedSpans.remove(span)) {
-      span.file.delete();
+      if (span.file != null) {
+        span.file.delete();
+      }
       return true;
     }
     return false;
@@ -212,5 +274,52 @@ import java.util.TreeSet;
         && key.equals(that.key)
         && cachedSpans.equals(that.cachedSpans)
         && metadata.equals(that.metadata);
+  }
+
+  private static final class Range {
+
+    /** The starting position of the range. */
+    public final long position;
+    /** The length of the range, or {@link C#LENGTH_UNSET} if unbounded. */
+    public final long length;
+
+    public Range(long position, long length) {
+      this.position = position;
+      this.length = length;
+    }
+
+    /**
+     * Returns whether this range fully contains the range specified by {@code otherPosition} and
+     * {@code otherLength}.
+     *
+     * @param otherPosition The position of the range to check.
+     * @param otherLength The length of the range to check, or {@link C#LENGTH_UNSET} if unbounded.
+     * @return Whether this range fully contains the specified range.
+     */
+    public boolean contains(long otherPosition, long otherLength) {
+      if (length == C.LENGTH_UNSET) {
+        return otherPosition >= position;
+      } else if (otherLength == C.LENGTH_UNSET) {
+        return false;
+      } else {
+        return position <= otherPosition && (otherPosition + otherLength) <= (position + length);
+      }
+    }
+
+    /**
+     * Returns whether this range intersects with the range specified by {@code otherPosition} and
+     * {@code otherLength}.
+     *
+     * @param otherPosition The position of the range to check.
+     * @param otherLength The length of the range to check, or {@link C#LENGTH_UNSET} if unbounded.
+     * @return Whether this range intersects with the specified range.
+     */
+    public boolean intersects(long otherPosition, long otherLength) {
+      if (position <= otherPosition) {
+        return length == C.LENGTH_UNSET || position + length > otherPosition;
+      } else {
+        return otherLength == C.LENGTH_UNSET || otherPosition + otherLength > position;
+      }
+    }
   }
 }
